moved method setConfigDefaultsInstance to TestUtils
# Conflicts:
#	java/core/src/test/java/com/redhat/rhn/testing/TestUtils.java
